table module, block module, refactored fig.
