+++ elifecleaner/__init__.py
@@ -1,7 +1,7 @@
 import logging
 
 
-__version__ = "0.51.0"
+__version__ = "0.52.0"
 
 
 LOGGER = logging.getLogger(__name__)
